refactoring: fix after code review
