+++ src/features/posts/ui/post/postForm/index.ts
@@ -1,1 +1,2 @@
 export * from './postForm'
+export * from './usePostForm'
